Add files via upload
This is a 12V lead acid battery charger using LM358 opamp. the charger can charger up to a 200AH battery by using a suited charging relay.
This circuit includes a battery level indicator using LM3914 configured in the Bar Mode to indicate the battery SOC to the user.
In the schematic we used potentiometers, but we intend to simulate and get suitable resistors for the final product to reduce the overall cost of the charger.
The battery level indications will be well showed in the final mechanical design.